feat(Stapel): split out ListNavigationLink
